test: cleanup test
